refactor: TimeControls is now a separate file
